Avoid SubscriptionManager#getUriForSubscriptionId calls with invalid subIds
It seems like the SIM card can take longer to load on certain legacy CDMA devices,
thus causing the subscription ID not to be valid during early telephony init.

Just skip calling SubscriptionController#setMccMnc or setCountryIso at that point,
so to avoid any subsequent calls to SubscriptionManager#getUriForSubscriptionId with
an invalid subId, which then makes the phone process crash. Once the SIM is loaded,
SubscriptionInfoUpdater will take care of updating the DB with the proper URI.

Change-Id: I8bd71127aecff9577bb0d3b9d7cab7a01a076e87